+++ src/main/kotlin/step3/IntRandom.kt
@@ -4,6 +4,7 @@ import java.util.Random
 
 interface IntRandom {
     fun next(): Int
+
     class Smart(private val random: Random, private val bound: Int = Int.MAX_VALUE) : IntRandom {
         override fun next() = random.nextInt(bound)
     }
